requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
click==6.6
configparser==3.5.0
cycler==0.10.0
decorator==4.0.9
Django==1.10
enum34==1.1.6
Flask==0.11.1
future==0.16.0
futures==3.0.5
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
geopy==1.11.0
google-gax==0.14.1
googleapis-common-protos==1.5.0
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.0.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
jws==0.1.3
MarkupSafe==0.23
matplotlib==1.5.1
nltk==3.2.1
nose==1.3.7
numpy==1.10.4
oauth2==1.9.0.post1
oauth2client==3.0.0
ply==3.8
praw==3.4.0
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycryptodome==3.4.3
pyowm==2.5.0
pyparsing==2.1.1
Pyrebase==3.0.24
python-dateutil==2.5.3
python-google-places==1.4.0
python-jwt==2.0.1
pytz==2016.4
pywsd==1.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.10.0
timezonefinder==2.0.1
update-checker==0.11
virtualenv==15.1.0
Werkzeug==0.11.11
yelp==1.0.2
